requirements.txt
pytest~=7.4.4
requests==2.31.0
wheel==0.38.1
aiohttp>=3.8.0,<3.9.0
psycopg2~=2.9.9
python-dotenv~=1.0.0
certifi~=2023.11.17
setuptools~=69.0.3